requirements.txt
streamlit>=1.35.0
google-generativeai>=0.8.1
pypdf>=4.2.0
python-docx>=1.1.2